+++ ichnaea/api/submit/submit_v2/schema.py
@@ -1,3 +1,5 @@
+import colander
+
 from ichnaea.api.schema import (
     OptionalIntNode,
     OptionalMappingSchema,
@@ -12,7 +14,7 @@ from ichnaea.api.submit.schema import (
 
 class CellTowerV2Schema(CellTowerSchema):
 
-    psc = OptionalIntNode()
+    psc = OptionalIntNode(internal_name='primaryScramblingCode')
 
 
 class CellTowersV2Schema(OptionalSequenceSchema):
@@ -22,7 +24,33 @@ class CellTowersV2Schema(OptionalSequenceSchema):
 
 class ReportV2Schema(PositionSchema, ReportSchema):
 
+    _position_fields = (
+        'latitude',
+        'longitude',
+        'accuracy',
+        'altitude',
+        'altitudeAccuracy',
+        'age',
+        'heading',
+        'pressure',
+        'speed',
+        'source',
+    )
+
     cellTowers = CellTowersV2Schema(missing=())
+
+    def deserialize(self, data):
+        data = super(ReportV2Schema, self).deserialize(data)
+        if data in (colander.drop, colander.null):
+            return data
+        position_data = {}
+        for field in self._position_fields:
+            if field in data:
+                position_data[field] = data[field]
+                del data[field]
+        if position_data:
+            data['position'] = position_data
+        return data
 
 
 class ReportsV2Schema(OptionalSequenceSchema):